# Conflicts:
#	src/mobius/models/qwen35.py

+++ src/mobius/rewrite_rules/_group_query_attention_test.py
@@ -321,14 +321,22 @@ class TestGroupQueryAttentionRules:
         assert counts.get("GroupQueryAttention", 0) == 0
         assert counts.get("Attention", 0) == num_layers
         for node in (node for node in model.graph if node.op_type == "QMoE"):
+            # router_probs (index 1) and router_weights (index 14) share
+            # QMoE's "T" type constraint with hidden_states, so they are
+            # cast back to the model dtype (FLOAT16 here) after being
+            # computed in FLOAT32 for routing-score numerical stability.
             assert node.inputs[1] is not None
-            assert node.inputs[1].dtype == ir.DataType.FLOAT
+            assert node.inputs[1].dtype == ir.DataType.FLOAT16
+            # fc1_scales/fc2_scales (indices 3, 6) must match the
+            # hidden_states/activation dtype (FLOAT16 here): QMoE's kernel
+            # registration requires T2 (scales) to exactly equal T
+            # (activation), so scales are no longer pinned to FLOAT32.
             assert node.inputs[3] is not None
-            assert node.inputs[3].dtype == ir.DataType.FLOAT
+            assert node.inputs[3].dtype == ir.DataType.FLOAT16
             assert node.inputs[6] is not None
-            assert node.inputs[6].dtype == ir.DataType.FLOAT
+            assert node.inputs[6].dtype == ir.DataType.FLOAT16
             assert node.inputs[14] is not None
-            assert node.inputs[14].dtype == ir.DataType.FLOAT
+            assert node.inputs[14].dtype == ir.DataType.FLOAT16
 
         input_names = [
             value.name
